LayoutTests/platform/ios-simulator/fast/events/data-transfer-files-attribute-identity-expected.txt

CONSOLE MESSAGE: line 17: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(['test.txt'])')

Test that the same object is returned for dataTransfer.files each time, as well as for File objects in the FileList.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
